refactor(metrics): unify process platform stats provider
